refactor: atualiza caminhos de importação e organiza formatação de colunas no DataTable
